Fixes QR decomposition problem and error in the Independent Weights graphs
The QR decomposition had a problem regarding the case where the position to be nulled and the position to become the "pivot" are both zeros, where it would still attempt to calculate sines and cossines and yeld a nan.
The Independent Weights graphs where calculated wrong: the subtraction was being done to an unexisting position of the grades matrix (column n_grades, it only goes up to and including n_grandes-1, the correct one).
Also, within the Independent Weights, the fourth weight was being attributed to the second report. That is false, it actually concerns the second test. The second report is the third weight.