Adds a new not-in-main function to calculate the weight configurations
Adds, in a new cpp file, the code of a new grade_calc function, which returns a matrix with the grade threshold on the first column and, for each line, the weight configurations. Also added a more generic way of calculating class averages and individual grades, so to compare and get the number of approved/final exam/fails more easily.